pyproject.toml
[tool.poetry]
name = "fitplate-ai"
version = "0.1.0"
description = "Async food image recognition and nutrition analysis web app using Streamlit"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
streamlit = "^1.33.0"
pillow = "^10.0.0"
aiohttp = "^3.9.0"
openai = "^1.30.1"
python-dotenv = "^1.0.0"
nest-asyncio = "^1.6.0"
pydantic = "^2.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
